Fixed behavior around retry on errors; added percentage and average speed calculations
Previously, on an error, the retry loop would end up reading the next
chunk of data from the file, so the resulting uploaded file would be
missing a chunk!  :(

Also added printing of the session ID and byte count to enable manual
resume in the case of a fatal error.