some documentation formatting fixes
1. File-scope doc comments (those starting with "//!") that refer to
types defined in the module need to have the link include the name of
the module, or rustdoc can't resolve the reference.
2. References to routes in another namespace were being generated with
multiple "super::" components in their path; instead, the human-visible
name should just be "namespace::route", and the link path needs to be
fixed to just "super::namespace::route".