avoid closures for default values in some cases
If a default value contains no parenthesis, consider it "trivial" and emit it directly in a .unwrap_or() instead of in a closure with .unwrap_or_else().

Otherwise, clippy warns of unnecessary_lazy_evaluations in these cases.

Unfortunately, Clippy has type information at its disposal and we do not, but this rough heuristic appears to be good enough.