Change default client from hyper 0.10 to ureq (#17)
The SDK currently uses Hyper 0.10 as the default HTTP client. This version is out of date (see #13) and unsupported, but the newer Hyper versions are async, and we're not ready to change dropbox-sdk to async yet.

Instead, this changes the default client to another blocking HTTP client crate, one which is actively maintained.

It also renames several types, replacing the name "hyper" with "default". By not specifying the implementation in the type names, we can change it in the future without renaming things again. See release notes for details.

This should not have any major impact on downstream crates beyond the renaming of types. ureq is a simple crate, does not do any threading of its own, and is configured here to use the system native TLS implementation, same as Hyper was. The dependency tree of dropbox-sdk is also significantly smaller after this change.