add python scripts for codegen and spec update
This replaces generate.sh with Python code so that it can be called from
the new spec update script.

Also library-ized update_manifest.py so it can be called from it as well.

The big feature of the spec update script is that it determines whether
the update is likely to be semver compatible or not, by doing the
following:
1. generate tests for the current spec
2. update the spec submodule
3. generate rust SDK code for the new spec
4. update the feature flags in the Cargo manifest
5. build and run the old tests against the new code
6. if they succeed, it's likely okay to just bump the patch version