better error type detection
For considering a type to be an error type, worthy of getting an impl of
`std::error::Error` and `std::fmt::Display`, we now use the conditions:

1. Is the type an error return type of any route? (this is new)
2. Is it an enum type with a name ending in 'Error'? (this is what we
had before)

The addition of condition 1 picks up two types that were missed before:

`files::UploadErrorWithProperties` (because it doesn't have an -Error suffix)

`team::TeamFolderListError` (because it's not a union, which honestly looks like a mistake, but it's in the spec so changing it would be an incompatible change)

`TeamFolderListError` is not worthy of writing additional code to handle
IMO, so it's going to just get the default crummy Display impl that
spits out a Debug repr, and it's not going to get a `source()` method
for `Error`, even though it could: it has a `TeamFolderAccessError`
inside it. Oh well.