give retry_after field in 429 responses a default value
Not all 429 errors have this field: `too_many_write_operations` isn't a time-based limit, so it doesn't include it. Currently, encountering this results in `Error::Json` instead of `Error::RateLimited` due to serialization failure in client_helpers.

We could change it to Optional<u32>, but that would be a breaking change; instead, default to 0.

It should hopefully be clear from a value of zero that the request can be retried immediately, as long as you're under the limit on concurrent operations.