examples/large-file-upload: ignore reuploaded block errors when resuming
Because we upload blocks in parallel, we may hit an error on a block but
still upload some blocks after that one. Then when resuming, we can hit
an error when the API complains that we already have data at an offset
we're trying to upload to. If we're resuming, assume that this is fine
and keep going.

(Ideally the API would see that the data in question matches the
already-uploaded block and not raise this error, but it doesn't...)