skip serializing struct fields that are None
Right now it writes something like:

    {"foobar": null}

which might work with the actual Python API code, but is unnecessarily
verbose, and doesn't work with our own deserializer, and round-tripping
is a desirable feature.

We could change the deserializers to accept `null` as a value for
`Option` types (maybe even in addition to this change) but the API
doesn't send JSON that way and IMO it's unnecessary to be able to read
that in.