Overhaul the OAuth2 helper code
Several things:
* implement basic support for short-lived tokens
* implement support for PKCE flow
* improve documentation in a few places

This is a breaking change due to renamed types, but migration should be
pretty straightforward: see the examples.

One thing remains: the Authorization should be integrated with the HTTP
clients, so that when an request encounters an expired OAuth2 token, it
can re-authorize using the refresh token and retry.