refactor: remove unbound variables for extraction of values from oauth2
Unbound variables were being used before being set to values from an
oauth2 token call result. This was slightly confusing because one of the
variables had already been declared. Now the code makes it clear which
variables are extracted from the result every time (expected) and which
variables are updated by the result if they existed in the result.