feat: add optional client_secret to AuthorizationState::InitialAuth
This changes the API somewhat significantly, but allows a refresh_token
returned by the PKCE OAuth2 flow to actually be used since it requires
a client secret to obtain a token, but not for authorizing the user with
the application. The client secret still remains hidden to an end user
during the OAuth2 flow.

It may still be confusing when and where to pass a client secret when
building an Authorization. Clearing this up could be done by changing
the OAuth2Type enum to require a client secret for the PKCE variant.