generated tests: emit assertions for implicit default values
When generating the OnlyRequiredFields tests, we need to check that the
implicit defaults (the omitted fields in the JSON) were deserialized to
the expected value, whether None or an other specified default value.

This means we have to still generate TestField instances for these
fields, since that's what's responsible for emitting the assertions.