async http client
This creates async HTTP client traits which mirror the existing sync
ones.

It also reimplements sync in terms of async, since we can use
now_or_never as an escape hatch to get back to sync land, but the
reverse is not possible.

One ugly wart is that there's no good escape hatch to convert the
sync-reader-in-async-wrapper response body back to a sync reader other
than an unsafe pointer cast.

No async routes generated yet; that's coming next.